requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyGithub==2.5.0
requests==2.32.3
python-dotenv==1.0.1
Jinja2==3.1.4
httpx==0.27.2

